requirements.txt
numpy==1.24.3
pandas==2.0.2
matplotlib==3.7.1
seaborn==0.12.2
ipython==8.13.2
jupyter==1.0.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
opencv-python==4.8.1.78
scikit-learn==1.3.0
tqdm==4.66.1
albumentations==1.3.1
timm==1.0.7
streamlit==1.16.0
altair==4.2.0